Fix file structure naming for components (ComponentTree and ComponentGraph
