+++ src/app/components/ComponentGraph/LinkControls.tsx
