+++ src/app/components/ComponentGraph/getLinkComponent.ts
