+++ src/app/components/ComponentTree3D/ComponentBox.tsx
