+++ src/app/containers/VisualContainer.tsx
@@ -3,8 +3,8 @@ import { ParentSize } from '@visx/responsive';
 import NavBar from '../components/NavBar/NavBar';
 import AtomNetwork from '../components/AtomNetwork/AtomNetwork';
 import StateDiff from '../components/StateDiff/StateDiff';
-import StateTree from '../components/StateTree/StateTree';
 import ComponentTree from '../components/ComponentTree/ComponentTree';
+import ComponentGraph from '../components/ComponentGraph/ComponentGraph';
 
 interface navType {
   [tabName: string]: JSX.Element;
@@ -15,11 +15,11 @@ function VisualContainer(): JSX.Element {
 
   const navLists: navType = {
     'State Diff': <StateDiff />,
-    'Component Tree': <StateTree />,
+    'Component Tree': <ComponentTree />,
     'Component Graph': (
       <ParentSize>
         {({ width, height }) => (
-          <ComponentTree  width={width} height={height} />
+          <ComponentGraph  width={width} height={height} />
         )}
       </ParentSize>
     ),
